Alternatives
Small changes to the main.py file.
Added a testing and testing2 files that have enhanced UI, where you can slide through all words in the output, not just one at a time.
This does not hit the main aim of the project (to create a complete tree system), but the remainder is arbitrary, and will involve looping to call the agent at each branch.

Key takeaways from the project:
1. tkinter - a basic interactive UI available for python visualisations.
2. Classes - more comfortable with classes now, including initiating them and their methods
3. OpenAI - more confident in using their API, and the arguments when doing so (for example logprobs and top_logprobs to analyse the LLMs thinking)
4. Debugging - Now have a thorough understanding of debugging python code, as I've realised its almost completely analogous to VBA debugging which I'm well practiced in. I will look into unit testing and other techniques moving forward.